sve: stack_round_01() could get the same result with t_round_05()
Fix the data size of each lane is 2048-bit. It couldn't be 512-bit.
Since we need to make it work on SVE-2048.

So loading value from stack could work now.

But only the first row of result from t_round_05() is same with
sample_round_05(). Other rows are all different.

Signed-off-by: Haojian Zhuang <haojian.zhuang@gmail.com>