crypto_m: test aes-256-cbc
Both aes-128-cbc and aes-256-cbc are sharing the same function in
crypto module. The only difference is the length of key. aes-128-cbc
needs 16-byte key and aes-256-cbc needs 32-byte key.

The encrypted result is similar as openssl command. But openssl
generates the encrypted result with 80-byte long, not 64-byte long.

Signed-off-by: Haojian Zhuang <haojian.zhuang@gmail.com>